expenses, users

+++ penny/db.go
@@ -12,6 +12,9 @@ func (ps *postgresStorage) init() error {
 	if err := ps.createIncomesTable(); err != nil {
 		return err
 	}
+	if err := ps.createExpensesTable(); err != nil {
+		return err
+	}
 	return nil
 }
 
@@ -20,14 +23,15 @@ func (ps *postgresStorage) createIncomesTable() error {
         id SERIAL PRIMARY KEY,
         title varchar(255) NOT NULL,
         amount REAL NOT NULL,
+        user_id INTEGER NOT NULL,
         created_at TIMESTAMPTZ DEFAULT(now())
     )`
 	_, err := ps.db.Query(query)
 	return err
 }
-func (ps *postgresStorage) getAllIncomes() ([]*Income, error) {
-	query := `SELECT id, title, amount, created_at FROM penny_incomes`
-	rows, err := ps.db.Query(query)
+func (ps *postgresStorage) getAllIncomes(userID int) ([]*Income, error) {
+	query := `SELECT id, title, amount, user_id, created_at FROM penny_incomes WHERE user_id = $1`
+	rows, err := ps.db.Query(query, userID)
 	if err != nil {
 		return nil, err
 	}
@@ -43,24 +47,23 @@ func (ps *postgresStorage) getAllIncomes() ([]*Income, error) {
 }
 func (ps *postgresStorage) createIncome(income *Income) (*Income, error) {
 	query := `INSERT INTO penny_incomes(
-        title, amount
+        title, amount, user_id
         ) VALUES (
-            $1, $2
-        ) RETURNING id, title, amount, created_at
+            $1, $2, $3
+        ) RETURNING id, title, amount, user_id, created_at
     `
-	rows, err := ps.db.Query(query, income.Title, income.Amount)
+	rows, err := ps.db.Query(query, income.Title, income.Amount, income.UserID)
 	if err != nil {
 		return nil, err
 	}
 	rows.Next()
 	return scanIntoIncomes(rows)
-
 }
-func (ps *postgresStorage) deleteIncomeByID(id int) (*Income, error) {
-	query := `DELETE FROM penny_incomes WHERE id = $1 RETURNING
-        id, title, amount, created_at
+func (ps *postgresStorage) deleteIncomeByID(id, userID int) (*Income, error) {
+	query := `DELETE FROM penny_incomes WHERE id = $1 AND user_id = $2 RETURNING
+        id, title, amount, user_id, created_at
     `
-	rows, err := ps.db.Query(query, id)
+	rows, err := ps.db.Query(query, id, userID)
 	if err != nil {
 		return nil, err
 	}
@@ -73,7 +76,72 @@ func scanIntoIncomes(rows *sql.Rows) (*Income, error) {
 		&income.ID,
 		&income.Title,
 		&income.Amount,
+		&income.UserID,
 		&income.CreatedAt,
 	)
 	return income, err
+}
+
+func (ps *postgresStorage) createExpensesTable() error {
+	query := `CREATE TABLE IF NOT EXISTS penny_expenses (
+        id SERIAL PRIMARY KEY,
+        title varchar(255) NOT NULL,
+        amount REAL NOT NULL,
+        user_id INTEGER NOT NULL,
+        created_at TIMESTAMPTZ DEFAULT(now())
+    )`
+	_, err := ps.db.Query(query)
+	return err
+}
+func (ps *postgresStorage) getAllExpenses(userID int) ([]*Expense, error) {
+	query := `SELECT id, title, amount, user_id, created_at FROM penny_expenses WHERE user_id = $1`
+	rows, err := ps.db.Query(query, userID)
+	if err != nil {
+		return nil, err
+	}
+	expenses := []*Expense{}
+	for rows.Next() {
+		expense, err := scanIntoExpenses(rows)
+		if err != nil {
+			return nil, err
+		}
+		expenses = append(expenses, expense)
+	}
+	return expenses, nil
+}
+func (ps *postgresStorage) createExpense(expense *Expense) (*Expense, error) {
+	query := `INSERT INTO penny_expenses(
+        title, amount, user_id
+        ) VALUES (
+            $1, $2, $3
+        ) RETURNING id, title, amount, user_id, created_at
+    `
+	rows, err := ps.db.Query(query, expense.Title, expense.Amount, expense.UserID)
+	if err != nil {
+		return nil, err
+	}
+	rows.Next()
+	return scanIntoExpenses(rows)
+}
+func (ps *postgresStorage) deleteExpenseByID(id, userID int) (*Expense, error) {
+	query := `DELETE FROM penny_expenses WHERE id = $1 AND user_id = $2 RETURNING
+        id, title, amount, user_id, created_at
+    `
+	rows, err := ps.db.Query(query, id, userID)
+	if err != nil {
+		return nil, err
+	}
+	rows.Next()
+	return scanIntoExpenses(rows)
+}
+func scanIntoExpenses(rows *sql.Rows) (*Expense, error) {
+	expense := &Expense{}
+	err := rows.Scan(
+		&expense.ID,
+		&expense.Title,
+		&expense.Amount,
+		&expense.UserID,
+		&expense.CreatedAt,
+	)
+	return expense, err
 }
